major graphics update Notes: PictureID (no longer an int) is a link to pictures, pictures will clean themself up when they are no longer referenced by a PictureID g_gr->get_no_picture() returns a PictureID to the 'invalid'/'empty' picture PicMod now is it's own type, it doesn't atck (PicMod_UI|PicMod_Game isn't valid anymore).
Look for graphic glitches



git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@4261 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/ai/defaultai.cc
@@ -1596,8 +1596,11 @@ bool DefaultAI::check_militarysites  (int32_t gametime)
 						// the destruction of the flag avoids that defaultAI will have
 						// too many unused roads - if needed the road will be rebuild
 						// directly.
-						if ((ms->maxSoldierCapacity() * 2) < bf.military_influence)
+						if
+						  (static_cast<int32_t>((ms->maxSoldierCapacity() * 2))
+							< bf.military_influence)
 							game().send_player_bulldoze(ms->base_flag());
+
 						// Else consider enhancing the building (if possible)
 						else {
 							// Do not have too many constructionsites
